refactor(gatsby): single function to load config and plugins
