+++ packages/gatsby/src/services/initialize.ts
@@ -11,21 +11,17 @@ import apiRunnerNode from "../utils/api-runner-node"
 import handleFlags from "../utils/handle-flags"
 import { getBrowsersList } from "../utils/browserslist"
 import { Store, AnyAction } from "redux"
-import { preferDefault } from "../bootstrap/prefer-default"
 import * as WorkerPool from "../utils/worker/pool"
 import { startPluginRunner } from "../redux/plugin-runner"
-import { loadPlugins } from "../bootstrap/load-plugins"
 import { store, emitter } from "../redux"
-import loadThemes from "../bootstrap/load-themes"
 import reporter from "gatsby-cli/lib/reporter"
-import { getConfigFile } from "../bootstrap/get-config-file"
 import { removeStaleJobs } from "../bootstrap/remove-stale-jobs"
 import { IPluginInfoOptions } from "../bootstrap/load-plugins/types"
-import { internalActions } from "../redux/actions"
 import { IGatsbyState } from "../redux/types"
 import { IBuildContext } from "./types"
 import availableFlags from "../utils/flags"
 import { detectLmdbStore } from "../datastore"
+import { loadConfigAndPlugins } from "../bootstrap/load-config-and-plugins"
 
 interface IPluginResolution {
   resolve: string
@@ -137,26 +133,17 @@ export async function initialize({
   emitter.on(`END_JOB`, onEndJob)
 
   // Try opening the site's gatsby-config.js file.
-  let activity = reporter.activityTimer(`open and validate gatsby-configs`, {
-    parentSpan,
-  })
-  activity.start()
-  const { configModule, configFilePath } = await getConfigFile(
-    program.directory,
-    `gatsby-config`
+  let activity = reporter.activityTimer(
+    `open and validate gatsby-configs, load plugins`,
+    {
+      parentSpan,
+    }
   )
-  let config = preferDefault(configModule)
+  activity.start()
 
-  // The root config cannot be exported as a function, only theme configs
-  if (typeof config === `function`) {
-    reporter.panic({
-      id: `10126`,
-      context: {
-        configName: `gatsby-config`,
-        path: program.directory,
-      },
-    })
-  }
+  const { config, flattenedPlugins } = await loadConfigAndPlugins({
+    siteDirectory: program.directory,
+  })
 
   // Setup flags
   if (config) {
@@ -206,22 +193,11 @@ export async function initialize({
   }
   detectLmdbStore()
 
-  // theme gatsby configs can be functions or objects
-  if (config) {
-    const plugins = await loadThemes(config, {
-      configFilePath,
-      rootDir: program.directory,
-    })
-    config = plugins.config
-  }
-
   if (config && config.polyfill) {
     reporter.warn(
       `Support for custom Promise polyfills has been removed in Gatsby v2. We only support Babel 7's new automatic polyfilling behavior.`
     )
   }
-
-  store.dispatch(internalActions.setSiteConfig(config))
 
   activity.end()
 
@@ -240,13 +216,6 @@ export async function initialize({
 
   // run stale jobs
   store.dispatch(removeStaleJobs(store.getState()))
-
-  activity = reporter.activityTimer(`load plugins`, {
-    parentSpan,
-  })
-  activity.start()
-  const flattenedPlugins = await loadPlugins(config, program.directory)
-  activity.end()
 
   // Multiple occurrences of the same name-version-pair can occur,
   // so we report an array of unique pairs
